+++ src/Dropdown/Dropdown.jsx
@@ -108,6 +108,11 @@ Dropdown.propTypes = {
   children: PropTypes.node.isRequired,
 
   /**
+   * Overrides the styles applied to the component.
+   */
+  classes: PropTypes.object,
+
+  /**
    * A boolean value indicating whether the dropdown is disabled.
    */
   disabled: PropTypes.bool,