+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* Add `classes` prop types
 * Add `component` prop type to `Button` component
 
 ## 1.6.4 (2018-12-12)